puddle/pyproject.toml
[tool.poetry]
name = "puddle"
version = "1.0"
description = "market project"
authors = ["DmitryAce <47420776+DmitryAce@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
attrs = "24.2.0"
certifi = "2024.7.4"
charset-normalizer = "3.3.2"
coreapi = "2.3.3"
coreschema = "0.0.4"
django = "5.1"
django-environ = "0.11.2"
django-filter = "24.3"
djangorestframework = "3.15.2"
drf-yasg = "1.21.7"
idna = "3.7"
importlib-resources = "6.4.3"
inflection = "0.5.1"
itypes = "1.2.0"
jinja2 = "3.1.4"
jsonschema = "4.23.0"
jsonschema-specifications = "2023.12.1"
markdown = "3.7"
markupsafe = "2.1.5"
openapi-codec = "1.3.2"
packaging = "24.1"
pillow = "10.4.0"
pytz = "2024.1"
pyyaml = "6.0.2"
referencing = "0.35.1"
requests = "2.32.3"
rpds-py = "0.20.0"
setuptools = "73.0.1"
simplejson = "3.19.3"
sqlparse = "0.5.1"
swagger-spec-validator = "3.0.4"
typing-extensions = "4.12.2"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
